+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## TBD
+
+### Enhancements
+
+- Default the `--project-root` to the current working directory for the `upload dsym` command. [148](https://github.com/bugsnag/bugsnag-cli/pull/148)
+
 ## 2.6.2 (2024-10-17)
 
 ### Fixes